+++ lambda/dynamic-fronts-fetcher/src/main.ts
@@ -1,52 +1,16 @@
 import * as process from 'node:process';
-import { Storage } from '@google-cloud/storage';
+import type { Storage } from '@google-cloud/storage';
 import { loadConfig } from './config';
-import { consumeReadable } from './consume-readable';
 import { getStorageClient } from './gcloud';
-import { IncomingDataRow, InvokeEvent } from './models';
-import { breakDownUrl, findMatchingFiles } from './url-handling';
+import { InvokeEvent } from './models';
+import {
+	breakDownUrl,
+	findMatchingFiles,
+	retrieveContent,
+} from './url-handling';
 
 function asyncDelay(timeout: number): Promise<void> {
 	return new Promise<void>((resolve) => setTimeout(resolve, timeout));
-}
-
-async function retrieveContent(
-	gcpBucket: string,
-	filePath: string,
-): Promise<IncomingDataRow[]> {
-	const storage = new Storage(); //just using app default; will probably need proper creds here
-	const b = storage.bucket(gcpBucket);
-	const f = b.file(filePath);
-
-	const content = await consumeReadable(f.createReadStream());
-	const objects = content.toString('utf-8').split('\n');
-
-	return objects
-		.map((str, ctr) => {
-			try {
-				return JSON.parse(str) as unknown;
-			} catch (err) {
-				console.warn(
-					`Unparseable content at line ${ctr} of ${gcpBucket}:${filePath} - '${str}'`,
-				);
-				return undefined;
-			}
-		})
-		.map((obj, ctr) => {
-			if (!obj) return undefined;
-			const result = IncomingDataRow.safeParse(obj);
-			if (result.success) {
-				return result.data;
-			} else {
-				console.warn(
-					`Data from line ${ctr} did not marshal: ${result.error.toString()}. Content was '${JSON.stringify(
-						obj,
-					)}'.`,
-				);
-				return undefined;
-			}
-		})
-		.filter((obj) => !!obj) as IncomingDataRow[];
 }
 
 let storage: Storage | null = null;
@@ -75,6 +39,13 @@ export const handler = async (eventRaw: unknown) => {
 	const files = await findMatchingFiles(storage, pathToScan);
 	console.log(`Got ${files.length} files`);
 
+	for (const f of files) {
+		const rows = await retrieveContent(f);
+		console.log(`${f.name} gave us ${rows.length} rows:`);
+		for (const r of rows) {
+			console.log(`  ${JSON.stringify(r)}`);
+		}
+	}
 	// const incomingData = await retrieveContent(event.gcpBucket, event.filePath);
 	// console.log(`Incoming data: ${JSON.stringify(incomingData)}`);
 };
